Renamed package ui.issuepanel.expanded to browserview
