+++ src/browserview/hideUI.js
