+++ src/browserview/mobileHideUI.js
